Download USGS water data
This is a function ('Download_NWIS_v2py.py') and a call file ('Test call
NWIS.py') that allow the user to download real-time surface water data
from the National Water Information System (NWIS) of the United States
Geological Survey (USGS).

The user provides a unique site ID number and a time period of data
required (start and end date) as strings. The obtained data is written
to a text file and saved in a specified directory for later, repeated
access.

The function is a simple web-crawler application.